smart-city-assistant/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
streamlit==1.37.1
requests==2.32.3
pinecone-client==4.1.2